requirements.txt
fitz==0.0.1.dev2
python-dateutil==2.9.0.post0
PyMuPDF==1.24.2
pytest==8.2.0